Added footer blocks
Added footer blocks along with a new image for the footer.  Put the config in the install area along with edits to the theme variables.  Removed a few unused twig files and more clean up to the sass files.  Fixed a bug with the back to top button displaying z-indexed below the subfooter layer.