Create add-to-project.yml
This automation provides the ability to automate issues and pull requests related to Github Projects (Beta). If the issue or pull request is not attached to a project, it is automatically added to the project and its desired column. In addition to status automation, this automation also supports custom field management.
https://github.com/leonsteinhaeuser/project-beta-automations